добавил темплейт в packageInfo.html теперь при загрузке в информации рендрерится нужный нам пакет, на который мы перешли по ссылке информация берется из url параметра
