+++ app/index.html
@@ -236,6 +236,7 @@
 				<defs><circle id="a" cx="10" cy="10" r="10"></circle><mask id="b" width="20" height="20" x="0" y="0" fill="#fff"><use xlink:href="#a"></use></mask></defs><g fill="none" fill-rule="evenodd" transform="translate(6 6)"><use stroke="#358ED7" stroke-width="4" mask="url(#b)" xlink:href="#a"></use><path fill="#358ED7" d="M10.887 5v3.984c0 .422-.02.837-.062 1.243a27.58 27.58 0 0 1-.167 1.292H9.494c-.07-.455-.125-.886-.167-1.292a12.36 12.36 0 0 1-.062-1.243V5h1.622zM9 14.123c0-.144.027-.28.08-.41a.982.982 0 0 1 .223-.335 1.097 1.097 0 0 1 .756-.306 1.054 1.054 0 0 1 .745.306 1.054 1.054 0 0 1 .306.745c0 .149-.028.287-.083.415a1.008 1.008 0 0 1-.557.55c-.13.053-.267.08-.411.08a1.089 1.089 0 0 1-.756-.3.99.99 0 0 1-.223-.33 1.063 1.063 0 0 1-.08-.415z"></path></g>
 			</symbol>
 		</svg>
+		<script type="module" src="js/getDate.js"></script>
 		<script type="module" src="js/drawRandomPackages.js"></script>
 		<script type="module" src="js/carousel.js"></script>
 		<script type="module" src="js/main.js"></script>
